The transmission of files now works properly
changed the fileInfoBuffer to be of size 512 and only the necessary bytes
are written when sending file information. This lets us retrieve the
fitting size one the other end so we dont need to worry about leftover
bytes from the previous transmission (since we only take the number of
bytes read out from the buffer).
Files are seperated by knowing the size of the file (which is now included
in the fileInfo part, as bytes 1-9).
Commented out the methods for only-files transmission in the
NetworkFileReceiver/Sender since those do not have the changes to run yet.
Will probably bring those up to date somewhere along the line.
Transformation of int/long to byteArrays and back is no handled by the
Utils class via static methods.
All functions for the Server are now enabled and are not run seperately
like before.
Several important parameters are now handed down from the main class so
that changing those can be done from one place. Will probably fetch those
from a ini file or something in the future.