enabled feature for nonemtpy client directory
Networksender and receicer now have a method to exchange a list of files
and their hashes. this enables the client to determine the missing/changed
files and only send those to the client if the client starts with a
nonempty directory (hence should save traffic)
The Networksender/receiver can now exchange a list of files as well as a
list of deltas. Those are in two seperate functions that use one common
function to do their job.